wordahead-backend/requirements.txt
# WordAhead Backend Requirements
# Compatible with Python 3.11

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Environment variables
python-dotenv==1.0.0

# HTTP requests
requests==2.31.0

# OpenAI (required by GP-TSM)
openai==1.6.1

# Production server
gunicorn==21.2.0

# Basic utilities
python-Levenshtein==0.21.0

# Optional: Uncomment if you need caching
# redis==5.0.1
